Datatypes used for both the preprocessing and only phases should be in common package. Also, rename TinyTableSBool to underline that it contains a share of a value.
