+++ src/main/java/dk/alexandra/fresco/suite/tinytables/storage/TinyTablesOnlineStorage.java
@@ -1,5 +1,7 @@
 package dk.alexandra.fresco.suite.tinytables.storage;
 
+import dk.alexandra.fresco.suite.tinytables.datatypes.TinyTable;
+
 public interface TinyTablesOnlineStorage {
 
 	public TinyTable getTinyTable(int id);
