+++ src/main/java/dk/alexandra/fresco/suite/tinytables/util/TinyTablesTripleGenerator.java
@@ -10,7 +10,7 @@ import dk.alexandra.fresco.framework.util.ot.OTReceiver;
 import dk.alexandra.fresco.framework.util.ot.OTSender;
 import dk.alexandra.fresco.framework.util.ot.datatypes.OTInput;
 import dk.alexandra.fresco.framework.util.ot.datatypes.OTSigma;
-import dk.alexandra.fresco.suite.tinytables.prepro.datatypes.TinyTablesTriple;
+import dk.alexandra.fresco.suite.tinytables.datatypes.TinyTablesTriple;
 
 public class TinyTablesTripleGenerator {
